Fix links to longer projects, or projects with dots
Previously, a bug existed because the project part of a link (`bar` in
`https://github.com/foo/bar`) was checked against the algorithm for usernames
instead of the algorithm for projects.  Usernames are stricter (such as not
allowing dots and being max 39 characters), which caused some links to not be
matched (such as those with dots, or those longer than 39 characters).

Closes GH-14.